chore(pennylane): remove dead code and consolidate tests
