+++ packages/devqubit-pennylane/src/devqubit_pennylane/circuits.py
@@ -521,48 +521,3 @@ def _compute_hashes(
             )
 
     return hash_circuit_pair(all_ops, total_nq, total_nc)
-
-
-def tapes_to_text(tapes: list[Any]) -> str:
-    """
-    Convert tapes to human-readable text diagrams.
-
-    Parameters
-    ----------
-    tapes : list
-        List of PennyLane tape objects.
-
-    Returns
-    -------
-    str
-        Combined text representation of all tapes.
-    """
-    parts: list[str] = []
-
-    for i, tape in enumerate(tapes):
-        if i > 0:
-            parts.append("")
-
-        parts.append(f"[Tape {i}]")
-
-        try:
-            # Try to use tape's string representation
-            parts.append(str(tape))
-        except Exception:
-            # Fallback: manual formatting
-            nq = _get_num_wires(tape)
-            parts.append(f"<Tape with {nq} wires>")
-
-            ops = getattr(tape, "operations", [])
-            for op in ops:
-                op_name = getattr(op, "name", type(op).__name__)
-                op_wires = list(getattr(op, "wires", []))
-                parts.append(f"  {op_name} @ {op_wires}")
-
-            meas = getattr(tape, "measurements", [])
-            for m in meas:
-                mtype = type(m).__name__
-                m_wires = list(getattr(m, "wires", []))
-                parts.append(f"  {mtype} @ {m_wires}")
-
-    return "\n".join(parts)
